tests/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
terraform-compliance>=1.3.0
ansible-lint>=6.0.0
molecule>=5.0.0
molecule-plugins[docker]>=23.0.0
docker>=6.0.0
requests>=2.28.0
azure-identity>=1.12.0
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=29.0.0
azure-mgmt-network>=23.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-monitor>=5.0.0
PyYAML>=6.0
jinja2>=3.1.0
